requirements.txt
Flask>=2.3.0
requests>=2.31.0
obsws-python>=1.6.2
python-dotenv>=1.0.0
pyinstaller>=6.1.0
